Quyen co cap nhat chuc nang them hang cho

+++ src/daos/Dao_VoucherGiamGia.java
@@ -100,7 +100,7 @@ public class Dao_VoucherGiamGia {
 			PreparedStatement preparedStatement = connection.prepareStatement(updateQuery);
 			preparedStatement.setString(1, voucher.getTenVoucher());
 			preparedStatement.setString(2,voucher.getMoTaChuongTrinh());
-			preparedStatement.setFloat(3, voucher.getPhanTramGiamTheoHoaDon());
+			preparedStatement.setFloat(3, Float.valueOf(voucher.getPhanTramGiamTheoHoaDon()).shortValue());
 			preparedStatement.setDate(4, java.sql.Date.valueOf(voucher.getNgayBatDau().toString()));
 			preparedStatement.setDate(5, java.sql.Date.valueOf(voucher.getNgayKetThuc().toString()));
 			preparedStatement.setBoolean(6, voucher.isTrangThai());
